refactor: tracker-py
